[T07-51] добавил адаптер для пользователя/поправил замечания
