+++ packages/client/src/app/hocs/store/withStore.tsx
@@ -1,6 +1,5 @@
 import React from "react"
-import {store} from "./store"
-
+import { store } from "./store"
 import { Provider } from "react-redux"
 
 const withStore = (component: () => React.ReactNode | JSX.Element) => () => (
