+++ packages/client/src/entities/auth/model/index.ts
@@ -1,1 +1,12 @@
-export * as AuthModel from './authSlice'
+import { reducer } from './authSlice'
+import { authErrorReasonSelector, isAuthUserSelector } from './selectors'
+
+const selectors = {
+  authErrorReasonSelector, 
+  isAuthUserSelector
+}
+
+export {
+  reducer,
+  selectors,
+}
