+++ packages/client/src/entities/auth/ui/AuthGuard.tsx
@@ -1,0 +1,32 @@
+import React, { useEffect } from 'react';
+import { useMeMutation } from '@/services/auth';
+import { Navigate } from 'react-router';
+import { CircularProgress } from '@mui/material';
+import { useAppSelector } from '@/app/hooks/redux';
+import { isAuthUserSelector } from '../model/selectors';
+
+interface OwnProps {
+  children: JSX.Element;
+}
+
+export const AuthGuard = ({ children }: OwnProps): JSX.Element | React.ReactElement => {
+  const [userQuery, { isError, isLoading }] = useMeMutation()
+
+  const isAuth = useAppSelector(isAuthUserSelector)
+
+  useEffect(() => {
+    if(!isAuth) {
+      userQuery({})
+    }
+  }, [isAuth])
+
+  if(isLoading) {
+    return <CircularProgress />
+  }
+
+  if(isError) {
+    return <Navigate to='/auth' replace={true} />
+  }
+
+  return children
+}
