+++ packages/client/src/pages/Signup/ui/Signup.tsx
@@ -2,10 +2,10 @@ import React, { FC, useCallback } from 'react'
 import Grid from '@mui/material/Grid'
 import useStyles from './styles'
 import { Form } from 'react-final-form'
-import { Button, Typography } from '@mui/material'
+import { Typography } from '@mui/material'
 import { Field } from 'react-final-form'
 import { SignupForm } from '@/shared/types/SignupForm'
-import { RoutePath } from '@/shared/config/routeConfig/routeConfig'
+import { RoutePath } from '@/shared/config'
 import { AppLink } from '@/shared/ui/AppLink/AppLink'
 import { TextFieldForm } from '@/shared/hocs/formHocs'
 import { useSignUpMutation } from '@/services/auth'
