+++ packages/client/src/shared/hocs/formHocs.ts
@@ -3,4 +3,4 @@ import { ComponentType } from 'react'
 import { FieldRenderProps } from 'react-final-form'
 import withFieldHOC from '@/app/hocs/form/withFieldHOC'
 
-export const TextFieldForm = withFieldHOC(TextField) as ComponentType<FieldRenderProps<any, HTMLElement, any> & TextFieldProps>
+export const TextFieldForm = withFieldHOC(TextField) as ComponentType<FieldRenderProps<unknown, HTMLElement, unknown> & TextFieldProps>
