+++ packages/client/src/shared/types/User.ts
@@ -4,6 +4,12 @@ export interface UserSignInDto {
 }
   
 export interface UserEntity {
+  fistName: string;
+  // необходимо вытащить из сваггера сущность юзера
+}
+
+  
+export interface UserServerEntity {
   fist_name: string;
   // необходимо вытащить из сваггера сущность юзера
 }
